add sequelize example add readme
