+++ x-ray-sample-server/index.html
@@ -10,6 +10,7 @@
   <p>To see a traced AWS SDK request, go <a href="/aws-sdk/">here</a>.</p>
   <p>To see a traced HTTP request, go <a href="/http-request/">here</a>.</p>
   <p>To see a traced MySQL query, go <a href="mysql">here</a>. Ensure you've configured a database as described in the README.</p>
+  <p>To see a traced sequelize connect, go <a href="/sequelize/">here</a>.</p>
   <p>To see a traced Dynamo Lambda S3 Request id 1, go <a href="/dynamo-lambda-s3/1">here</a>.</p>
   <p>To see a traced Dynamo Lambda S3 Request id 2, go <a href="/dynamo-lambda-s3/2">here</a>.</p>
 </body>
