+++ x-ray-sample-server/package.json
@@ -13,6 +13,7 @@
   "devDependencies": {
     "@types/express": "^4.17.12",
     "@types/node": "^16.0.0",
+    "@types/validator": "^13.6.3",
     "typescript": "^4.3.5"
   },
   "dependencies": {
@@ -20,6 +21,8 @@
     "aws-sdk": "^2.939.0",
     "aws-xray-sdk": "^3.3.3",
     "express": "^4.17.1",
-    "mysql": "^2.18.1"
+    "mysql": "^2.18.1",
+    "mysql2": "^2.2.5",
+    "sequelize": "^6.6.5"
   }
 }
